requirements.txt
python-django
django-bootstrap5
django-admin-interface
django-colorfield
python-twisted
python-slugify
django-debug-toolbar
python-yaml
python-channels
python-daphne
paramiko
python-pycryptodome
channels_redis
python-msgpack
python-redis
